+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 ### Additions
 
 - `readonly` and `cancel` pseudoclasses added to type safe CSS
-
+- Added add/remove/toggle class for Tab
 - ContextMenu.radiomenuitem (https://github.com/edvin/tornadofx/issues/646)
 - mutableList can now be bound to an ObservableMap
 ## [1.7.15]